[#102] attribute 파싱 로직을 별도 파일로 분리
