+++ rupring_macro/src/parse.rs
@@ -1,8 +1,8 @@
-use std::collections::HashMap;
-
 use proc_macro::{TokenStream, TokenTree};
 use quote::ToTokens;
 use syn::{ItemFn, ItemStruct};
+
+use crate::attribute::{self, AnnotatedParameter};
 
 // Find the structure name immediately to the right of the struct keyword.
 pub(crate) fn find_struct_name(struct_ast: &ItemStruct) -> String {
@@ -49,144 +49,6 @@ pub(crate) fn find_function_return_type(item: TokenStream) -> String {
     return return_type;
 }
 
-#[derive(Debug, PartialEq, Clone)]
-pub enum AttributeValue {
-    ListOfString(Vec<String>),
-    String(String),
-}
-
-impl AttributeValue {
-    pub fn as_string(&self) -> String {
-        match self {
-            AttributeValue::String(value) => value.clone(),
-            AttributeValue::ListOfString(value) => value.join(","),
-        }
-    }
-}
-
-pub(crate) fn extract_additional_attributes(
-    item: TokenStream,
-) -> (TokenStream, HashMap<String, AttributeValue>) {
-    let mut map = HashMap::new();
-
-    let mut code_without_attributes: Vec<proc_macro::TokenTree> = vec![];
-    let mut iter = item.into_iter();
-    let mut done = false;
-
-    while let Some(tree) = iter.next() {
-        if done {
-            code_without_attributes.push(tree);
-            continue;
-        }
-
-        match tree {
-            proc_macro::TokenTree::Punct(ref punct) => {
-                if punct.to_string().as_str() == "#" {
-                    // [Key1 = value1, key2, value2, ...] 형태의 attribute를 파싱해서 map에 할당한다.
-                    if let Some(group) = iter.next() {
-                        if let proc_macro::TokenTree::Group(group) = group {
-                            let attributes = parse_attribute(group.stream(), true);
-
-                            for (key, value) in attributes {
-                                map.insert(key, value);
-                            }
-                        }
-                    }
-                } else {
-                    code_without_attributes.push(tree);
-                    done = true;
-                }
-            }
-            _ => {
-                code_without_attributes.push(tree);
-                done = true;
-            }
-        }
-    }
-
-    (code_without_attributes.into_iter().collect(), map)
-}
-
-// controllers = [HomeController {}], modules = [] => HashMap<String, AttributeValue>
-pub(crate) fn parse_attribute(
-    item: TokenStream,
-    with_alias: bool,
-) -> HashMap<String, AttributeValue> {
-    let mut tokens = item.into_iter();
-    let mut attribute_map = HashMap::new();
-
-    let mut attribute_name = None;
-    while let Some(token) = tokens.next() {
-        let token_string = token.to_string();
-
-        if token_string == "=" {
-            let mut attribute_name = attribute_name.clone().unwrap();
-            let mut attribute_value = tokens
-                .next()
-                .expect("key/value pair does not match")
-                .to_string();
-
-            if with_alias {
-                if attribute_name == "path" || attribute_name == "Path" {
-                    attribute_name = "PathVariable".into();
-                }
-
-                if attribute_name == "desc"
-                    || attribute_name == "Desc"
-                    || attribute_name == "description"
-                {
-                    attribute_name = "Description".into();
-                }
-            }
-
-            let attribute_value = if attribute_value.starts_with("[") {
-                let attribute_value = attribute_value
-                    .trim_start_matches("[")
-                    .trim_end_matches("]")
-                    .to_string();
-
-                let attribute_value = attribute_value
-                    .split(",")
-                    .map(|s| s.trim().to_string())
-                    .collect::<Vec<String>>();
-
-                // filter empty string
-                let attribute_value = attribute_value
-                    .into_iter()
-                    .filter(|s| s.len() > 0)
-                    .collect::<Vec<String>>();
-
-                AttributeValue::ListOfString(attribute_value)
-            } else {
-                while let Some(token) = tokens.next() {
-                    let token_string = token.to_string();
-
-                    if token_string == "," || token_string == "=" || token_string == ")" {
-                        break;
-                    }
-
-                    attribute_value.push_str(&token_string);
-                }
-
-                AttributeValue::String(attribute_value)
-            };
-
-            attribute_map.insert(attribute_name, attribute_value);
-        } else {
-            attribute_name = Some(token_string);
-        }
-    }
-
-    return attribute_map;
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct AnnotatedParameter {
-    pub(crate) attributes: HashMap<String, AttributeValue>,
-    pub(crate) name: String,
-    pub(crate) type_: String,
-}
-
 // 1. 타입 일관성을 위해 Request와 Response 매개변수가 존재하지 않는다면 강제로 추가합니다.
 // 2. 어노테이션이 붙은 특수한 파라미터를 제거해서 반환합니다.
 pub(crate) fn manipulate_route_function_parameters(
@@ -229,7 +91,7 @@ pub(crate) fn manipulate_route_function_parameters(
 
                                         let mut attributes =
                                             if let proc_macro::TokenTree::Group(group) = group {
-                                                parse_attribute(group.stream(), true)
+                                                attribute::parse_attribute(group.stream(), true)
                                             } else {
                                                 panic!(
                                                     "invalid annotation parameter (group expected)"
@@ -246,7 +108,10 @@ pub(crate) fn manipulate_route_function_parameters(
                                                     if let proc_macro::TokenTree::Group(group) =
                                                         group
                                                     {
-                                                        parse_attribute(group.stream(), true)
+                                                        attribute::parse_attribute(
+                                                            group.stream(),
+                                                            true,
+                                                        )
                                                     } else {
                                                         panic!(
                                                             "invalid annotation parameter (group expected)"
